+++ miden-vm/tests/integration/operations/sys_ops.rs
@@ -83,6 +83,6 @@ fn assert_eq_fail() {
 #[test]
 fn emit() {
     let mut test = build_op_test!("emit.42", &[0, 0, 0, 0]);
-    test.add_event_handler(42, |_| Ok(()));
+    test.add_event_handler(42, |_| Ok(Vec::new()));
     test.prove_and_verify(vec![], false);
 }